Initial upload
This code looks for a file named "passwd" in the current directory, opens it and counts the occurrences of each login shell, then prints the results, at least that's what it was meant to do. Obviously, as a rank rust beginner, my rust fu is weak, and I'm having trouble trying to overcome the fatal compiler errors.Suggestions welcome.